requirements.txt
pandas
numpy
requests
python-dotenv
SQLAlchemy
